refactor: move transaction history components into one spot
